+++ docs/docs/99-api/examples/blankspaces_demo/blankspaces_demo.js
@@ -2,7 +2,7 @@
  * This file demonstrates the ability to leave "blanks" when the 'automated.md' option is on.
  *
  * To define a blank box, you specify it as an object in the array (the classic array of objects) with three attributes:
- * - type: This must be equal to "BLANK"
+ * - type: This must be equal to ".blank"
  * - width: the desired width of the blank box (I say box but in reality there aren't any borders)
  * - height: the desired height of the blank box
  *
@@ -21,19 +21,17 @@ const WIDTH = 1300;
 
 const listOfObjs = [
     {
-        isClass: true,
+        type: ".frame",
         name: "__main__",
         id: null,
         value: { lst1: 82, lst2: 84, p: 99, d: 10, t: 11 },
-        stack_frame: true,
     },
-    { type: "BLANK", width: 100, height: 200, stack_frame: true },
+    { type: ".frame", name: ".blank", width: 100, height: 200 },
     {
-        isClass: true,
+        type: ".frame",
         name: "func",
         id: null,
         value: { age: 12, name: 17 },
-        stack_frame: true,
     },
     { type: "list", id: 82, value: [19, 43, 28, 49] },
     {
@@ -43,10 +41,10 @@ const listOfObjs = [
         show_indexes: true,
     },
     { type: "int", id: 19, value: 1969 },
-    { type: "BLANK", width: 100, height: 200 },
+    { type: ".blank", width: 100, height: 200 },
     { type: "bool", id: 32, value: true },
     { type: "str", id: 43, value: "David is cool" },
-    { type: "BLANK", width: 200, height: 150 },
+    { type: ".blank", width: 200, height: 150 },
     { type: "tuple", id: 11, value: [82, 76] },
 ];