+++ docs/docs/99-api/examples/style_demo/presets_demo.js
@@ -12,22 +12,20 @@ const WIDTH = 1300;
 
 const objs = [
     {
-        isClass: true,
+        type: ".frame",
         x: 25,
         y: 200,
         name: "__main__",
         id: null,
         value: { lst1: 82, lst2: 84, p: 99, d: 10, t: 11 },
-        stack_frame: true,
     },
     {
-        isClass: true,
+        type: ".class",
         x: 350,
         y: 10,
         name: "tuple",
         id: 99,
         value: { age: 12, name: 17 },
-        stack_frame: false,
         style: ["highlight_type"],
     },
     {